Refactoring response services&controller of items

+++ src/users/users.controller.ts
@@ -3,6 +3,8 @@ import {
     Controller,
     Delete,
     Get,
+    HttpException,
+    HttpStatus,
     Param,
     Patch,
     UseGuards,
@@ -14,12 +16,12 @@ import { UsersService } from './users.service';
 import { AuthGuard } from 'src/auth/guards/auth.guard';
 import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
 import { User } from '@prisma/client';
-import { CreateUserDto } from './dto/create-user.dto';
 
 
-@ApiBearerAuth()
+
+@ApiBearerAuth() // pr dire que les routes st protétgées
 @ApiTags("Users")
-//@UseGuards(AuthGuard) a remettre lors de la phase sécurisation de l'appli
+//@UseGuards(AuthGuard)
 @Controller('users')
 export class UsersController {
     constructor(private readonly usersService: UsersService) { }
@@ -36,17 +38,29 @@ export class UsersController {
 
     @Get(':id')
     async findOne(@Param('id') id: string): Promise<User> {
+        const user = await this.usersService.findOneById(id)
+        if (!user ) {
+            throw new HttpException("User not found", HttpStatus.NOT_FOUND);
+        }
         return await this.usersService.findOneById(id);
     }
 
     @Patch(':id')
     async update(@Param('id') id: string, @Body() data: UpdateUserDto): Promise<User> {
+        const user = await this.usersService.findOneById(id)
+        if (!user ) {
+            throw new HttpException("User not found", HttpStatus.NOT_FOUND);
+        }
         return await this.usersService.update(id, data);
     }
 
-    /* @Delete(':id')
-     remove(@Param('id') id: string, @Body() data: CreateUserDto) {
-      
-         return this.usersService.remove(id);
-     }*/
+    @Delete(':id')
+    async delete(@Param('id') id: string){
+        // if user.role (admin ..... a revoir)
+        const user = await this.usersService.findOneById(id)
+        if (!user ) {
+            throw new HttpException("User not found", HttpStatus.NOT_FOUND);
+        }
+        return await this.usersService.remove(id);
+    }
 }
